refactor: crate io mixin for docvec with tensor type
